refactor: ViewModels packaging changed
