+++ app/src/main/java/com/smmousavi/reactiveflow/ui/viewmodel/MainActivityViewModel.kt
@@ -1,4 +1,4 @@
-package com.smmousavi.reactiveflow.ui.activity
+package com.smmousavi.reactiveflow.ui.viewmodel
 
 import android.app.Application
 import androidx.compose.runtime.mutableStateOf
